requirements.txt
python-dotenv
google-adk
litellm
mcp
beautifulsoup4
lxml
